requirements.txt
Flask==2.3.2
facebook-sdk==3.1.0
gunicorn==20.1.0